Fix #copy_to_tempfile with nested basename
If a nested basename is passed to #copy_to_tempfile, Tempfile will try
to create a file in that directory, but that nested directory normally
won't exist, so an Errno::ENOENT error is thrown. It's better that we
always use "down" as a basename, which is also nicely distinguishable
from the "open-uri" tempfile that the open-uri library generates.