Move old open-uri file instead of copying it
When we download the file via open-uri, and want to correct the file
extension, we move it if it's a tempfile instead of copying it. This
difference is especially important when downloading large files. We also
reopen the Tempfile just in case, because aws-sdk was having problems
with a similar situation in ImageProcessing gem.