Remove decoding-encoding of the given URL
This was originally done so that users can pass in unencoded URLs.
However, it causes problems because `URI.encode` doesn't encode "[]" and
encodes "#" symbol:

  URI.encode("http://example.com/foo[1].jpg") #=> "http://example.com/foo[1].jpg"
  URI.encode("http://example.com/#foo")       #=> "http://example.com/%23foo"

More specifically, `URI.encode` and `URI.decode` aren't inverse
functions to each other. Furthermore, both `URI.encode` and `URI.decode`
are deprecated.

The correct behaviour is to require the given URL to be properly
encoded. This might be a breaking change, but honestly in context of
Shrine I can't think of a scenario where the user would copy-paste an
unencoded URL.

Fixes janko-m/shrine#38