Use human-readable error messages
When raising Down::TooLarge, we include the actual error message and the
maximum file size, which is suitable for displaying to the end user.

When raising Down::Error, often the error messages raised by Ruby would
be too complicated for the end user, so we rather use the simple "file
not found" message. Initially we included the original error in the
message, but on Ruby 2.1 or later people can just use `Exception#cause`
to retrieve the original error.