Add #stream for streaming chunks of remote files
The open-uri standard library internally streams files in chunks,
however it doesn't have any option to yield these chunks. It would be
really useful to make a PR for that. In the meanwhile, we try to make
our own implementation of streaming, taking just a small subset of
open-uri's implementation which should work in most situations, but may
not cover some edge cases.

A use case for this method is being able to stream a remote file as
response in an HTTP endpoint. If the file is first downloaded and then
given as the response, this is much slower.