Improve extracting original filename
Firstly, we do CGI.unescape only after we extract the basename, since
theoretically the filename could contain slashes (which were escaped in
the URL), and in that case we don't want to return an incomplete
filename.

Secondly, we switch from File.basename to explicitly splitting on "/"
characters, as I think File.basename wouldn't work on Windows because it
likely would be looking only for backslashes.