Add Down.open for IO-like streaming
In Shrine I needed a way to open an IO-like object around a remote file,
which will internally download the file on-the-fly as it is read, but
only how much is needed.

For example, when you want to determine the MIME type of the remote
file, you only need to download the beginning of the file, because this
is where the information about the MIME type is written.