Mark response as read before terminating download
The response is automatically marked as read (@read is set to true) when
`Response#read_body` gets to the end. However, the goal of `Down.open`
was to be able to terminate download at any point. This means that
`Response#read_body` method doesn't finish and doesn't set @read to
true, but then unfortunately the `#request_get` block still reads the
whole response.

So we need to tell net/http that we've read the response (all of the
response that we needed, anyway), which means that net/http will now
properly terminate the connection without reading the rest of the
response.