Work around Fiber errors with chunked responses
When response is returned as "Transfer-Encoding: chunked", attempting to
use Fibers (via Enumerators) raises an error:

  FiberError: fiber called across stack rewinding barrier

I haven't determined the cause of this error, but it is due to Net::HTTP
having different implementation of reading the response body. Since we
need Fibers for partial downloads, and Net::HTTP doesn't let us with
chunked responses, the only thing we can do is to download the whole
response.