Don't write chunks to file on ChunkedIO#each_chunk
We want Down::ChunkedIO#each_chunk to be a way for users to stream
response body directly, without additional IO overhead and disk space
usage of writing into a file.

Down::ChunkedIO#each_chunk is used in Shrine, in the download_endpoint
plugin. Having chunks downloaded to disk means the user needs to
consider whether they have enough disk storage during the download, and
the downloaded file is never even used.

This is a backwards incompatible change for anyone relying on the
Down::ChunkedIO object be rewindable after #each_chunk finishes, but
it's highly unlikely someone wants to yield chunks and use IO methods on
the same Down::ChunkedIO object.