Switch from Webmock stubs to a real HTTP server
Webmock has limitations, for example we cannot test streaming with it.
Moreover, we won't be able to use it for future wget or HTTP.rb
backends.

Fortunately, there is a service for testing HTTP requests, httpbin.org,
which is open source and can be run locally. This commit switches from
Webmock to httpbin.