Reraise only a subset of exceptions as Down::NotFound
It's not nice user experience to wrap all exceptions into
Down::NotFound, there might be some exceptions that are really important
for the user to know about.

So we rescue only a subset of exceptions that we think are safe which
will be reraised as Down::NotFound, and the rest of the errors are
propagated.