Add HTTP.rb backend
HTTP.rb is a really stable and awesome HTTP client library (unlike
Net::HTTP), so there are many benefits in using HTTP.rb as the Down
backend. Some of these benefits include:

* Correct URI parsing with Addressable::URI
* Support for streaming downloads (#download can now reuse #open)
* Support for SSL
* HTTP client builder API for setting default options
* Support for persistent connections
* Support for auto-inflating compressed response bodies

The HTTP.rb backend tries to include as much of the functionality that the
Net::HTTP backend had as possible.