Add :encoding option to Down::ChunkedIO.new
The returned content from Down::ChunkedIO should have consistent
encoding. Currently the content will be returned in UTF-8 encoding when
read the first time (since the buffer string is in UTF-8 encoding), and
in binary encoding when read from the cache.

We fix this inconsistency and force binary encoding. We also add the
:encoding option for overriding the default binary encoding.