Run kennethreitz/httpbin as a container
Currently we're installing and running httpbin as a Python application.
One limitation is that it requires users to have Python and Pip
installed. But second, primary limitation, is that we cannot run it on
Travis CI, because Travis supports only one language.

The initial solution to this problem was to hit http://httpbin.org
directly on Travis. That still had good performance, but unfortunately
the requests would often fail with a "503 Service Unavailable" error.
And that makes sense, because likely more people are doing the same
thing, and it's not sustainable.

Luckily, httpbin is also hosted on the Docker registry, so we can just
use Docker to run httpbin. Since Travis CI supports Docker, this way we
can have the exact same setup locally and on Travis CI.