Return an Integer in :status field with HTTP.rb
HTTP::Response#status returns an HTTP::Response::Status instance, which
is a delegator class to the underlying Integer. We want the :status
field to return a pure Integer, for consistency with the NetHttp
backend, and we already have the :response object if we really want to
retrieve a HTTP::Response::Status object.