In Down::ChunkedIO lazily create the Tempfile
In Shrine's download_endpoint plugin we're using ChunkedIO#each_chunk to
stream the file into the response body. The benefit of
ChunkedIO#each_chunk is that it doesn't cache the retrieved chunks into
a Tempfile, even when it's rewindable, so we aren't paying unnecessary
IO calls and disk usage for nothing. We couldn't just set :rewindable to
false there, because we rely only on the general storage abstraction,
that doesn't necessarily have to return a Down::ChunkedIO, nor does that
storage method accept additional options in the first place.

So, we have the benefit of ChunkedIO#each_chunk not writing the
retrieved chunked into the Tempfile, but the Tempfile is still being
created on initialization. This creates an additional file descriptor
for the whole duration of download that is not used. Since operating
systems have limited number of file descriptors, we should strive not to
create any unnecessary file descriptors.

So we move creation of the Tempfile into a method that will create the
Tempfile only when it's called. Since Chunked#each_chunk doesn't call
that method, the Tempfile and its file descriptor won't be created.