Make Down::ChunkedIO#inspect prettier
Down::ChunkedIO#inspect by default shows all instance variables that are
being used, which is too much information, because many of them are
internal. So we override Down::ChunkedIO#inspect to show only meaningful
information, and add "(closed)" when it's closed, to match #inspect output
of built-in Ruby IO objects.