Add Down::ChunkedIO#readpartial
The semantics of Down::ChunkedIO#readpartial is the same as in all IO
objects: it first returns any data that's already retrieved (up to the
specified maxlen), and only retrieves new data when retrieved data has
been emptied.

Down::ChunkedIO#readpartial is more memory efficient than using
Down::ChunkedIO#read, because in most cases chunks won't be split, and
splitting allocates new strings.

Consider a Down::ChunkedIO object initialized by three chunks, each 10KB
in size. If we copy this object using `IO.copy_stream` (which reads in
16KB at a time) without #readpartial, we would allocate two additional
strings, because the second chunk would be split into two.

  io = Down::ChunkedIO.new(chunks: chunks)
  io.read(16*1024).bytesize #=> 16*1024
  io.read(16*1024).bytesize #=> 14*1024

With #readpartial, on the other hand, which IO.copy_stream would
automatically pick up by the way, the chunks wouldn't be split.

  io = Down::ChunkedIO.new(chunks: chunks)
  io.read(16*1024).bytesize #=> 10*1024
  io.read(16*1024).bytesize #=> 10*1024
  io.read(16*1024).bytesize #=> 10*1024

The only case way chunks will be split is if maxlen is smaller than size
of chunks. However, considering that Shrine, IO.copy_stream, HTTP.rb
etc. all use chunk size of 16 KB, because it happens to be the buffer
size of a TCP socket, this won't happen in most cases because
Down::ChunkedIO is mostly used for HTTP downloads.