Halve string allocations in ChunkedIO#readpartial
Previously Down::ChunkedIO#readpartial without a buffer would allocate
a new string for each chunk it reads. But this is not necessary, we can
just return the retrieved chunk directly, which is exactly what we do in
this refactoring. This halves the number of string allocations when
using ChunkedeIO#readpartial without a buffer string, because we're not
allocating any new strings.