Explicitly require and assign NetHttp backend
I think it's too dynamic that the NetHttp backend is assigned on-the-fly
before first Down.download or Down.open is called. Yes, now people
always have to require the NetHttp backend, even if they switch to
another backend globally, but that's life.

This also fixes a case where someone was relying on the standard
libraries that the NetHttp backend required in Down 3.x or lower (e.g.
open-uri), and now on 4.0 with NetHttp backend being required
on-the-fly, so these standard libraries are not available at first.

Changing the global backend will be such a rare thing that it's worth
the performance tradeoff of requiring some things which won't be used.

Fixes #15