Remove the Down.backend= setter
This was originally done so that users don't have to remember whether
they need to use reader with an argument or a setter. However, these two
versions don't have the same behaviour, the setter doesn't accept
symbols. So they're *not* interchangeable, and that can cause confusion.

Therefore we remove this setter, it wasn't even documented anyway.