Rescue and re-raise timeout errors in more places
This library makes a promise to raise Down::Error exceptions whenever
known errors occur, but currently we're not watching out for a few
places where a timeout error can be raised, so it's still possible for
users to get raw Timeout::Error exceptions.

We fix this by rescuing timeout errors when instantiating a
Down::ChunkedIO (because Down::ChunkedIO reads the first chunk of the
body when it's instantiated), and when reading body (because there can
be a timeout while reading from the socket at any time).

Fixes #16