Improve ChunkedIO memory usage with buffer string
When a buffer string is used, the string objects that we retrieved from
the Enumerator won't be returned to the user, so we can deallocate them
with String#clear.

I used MemoryProfiler to profile a 18MB download, and the allocated
memory went from 20MB down to only 1MB, which is 20x improvement.