Improve memory usage of Down::Http.download
Down::ChunkedIO now deallocates chunks returned by the Enumerator when a
buffer string is used, which can improve memory usage up to 20x times.
Since in Down::Http.download we're not retaining the retrieved string
objects, we modify it to use a buffer string.

I used MemoryProfiler to profile a 18MB download, and the allocated
memory went from 20MB down to only 1MB, which is an improvement of 20x.