Add :method to Down::Http for specifying the request method
By default Down::Http will make GET requests, and we add the :method
option for overriding the request method.

One use case for this is the ImageOptim.com service, which accepts POST
requests and returns processed files.

open-uri doesn't support overriding the request method, so we can't
add the :method option to Down::NetHttp. On the other hand, Down::Wget
already supports :method because "--method" is one of wget's options.