Raise clear ResponseError on invalid redirect URI
When redirect URI cannot be parsed, Down::NetHttp#download currently
raises ResponseError that just includes the status code and description
in the error message, while Down::NetHttp#open raises a raw
URI::InvalidURIError. This is not explanatory, as in the #download case
it simply looks like Down failed to follow the redirect, while in the
\#open case we don't know where the invalid URL came from.

This commit makes both Down::NetHttp#download and Down::NetHttp#open
raise a ResponseError exception with explanatory message.

Closes #20