Modify previous commit to raise ResponseError
We already return ResponseError when redirect URL is considered invalid,
so we want to keep the consistency.

We currently can't do the same with Down::Http, as it doesn't allow us
to find out whether the original URL was invalid or one of the redirect
URLs. So at the moment Down::Http will raise Down::InvalidUrl in this
case.