Make internal Tempfile of Down::ChunkedIO inaccessible to outside programs
When opening a Down::ChunkedIO to a remote file, the user might want to
directly stream the contents to some other destination, for example an
AWS S3 bucket. Since Down::ChunkedIO is by default caching the retrieved
content to disk for rewindability, other programs can theoretically
retrieve the downloaded file content before the Tempfile is deleted, by
searching for filesystem entries.

This is not ideal in terms of security. To combat that, other libraries
like Rack (Rack::RewindableInput) and Unicorn (Unicorn::TeeInput) use a
technique where the file permissions are immediately stripped to
completely prohibit read & write, and in case of POSIX filesystems the
file entry is even removed from the filesystem. This ensures no one can
access the file other than the Ruby process that's holding the file
descriptor.

So, we copy the same technique here.