Make ChunkedIO#read & #readpartial with length always return strings in binary encoding
IO#read and IO#readpartial, when given a length, will always return
strings in binary encoding, even if the IO encoding was UTF-8. So we
mimic that behaviour in Down::ChunkedIO#read and #readpartial, by
forcing to binary encoding when length was given.

This fixes the issue where the result of ChunkedIO#read was a string
where #valid_encoding? returned false, because it was an UTF-8 string
with binary characters. This was caused by String#byteslice, which keeps
the UTF-8 encoding even if the characters are binary.

Concerns #30