Make ChunkedIO#read(partial) work with frozen chunks
Normally we were careful not to mutate the chunks, but we forgot that
String#force_encoding is destructive. So we work around that by
duplicating the chunk if we need to change its encoding, and update the
test suite to use frozen string literals in general.

This shouldn't have a memory impact in most common cases, which is
working with binary data. But it will in cases when `#read(partial)`
needs to return an encoding that's different from :encoding and/or
chunks' encoding. That's a bit unfortunate, but it's more important to
return results in correct encoding.

Fixes #32