Various Down::ChunkedIO improvements
Instead of trying to optimize every path, we make sure to optimize only
the most important one, which is the one with an output buffer. This
makes the code a bit easier to reason about.

We also make sure that we're dealing with binary encoding all the way
until we need to return the result, and only then force to specified
encoding if it's different than binary.

These changes allow us to easily handle frozen chunks, which fixes #32.

While we're at it, we also greatly improve ChunkedIO#gets memory
allocation.