Repurpose Down::NotFound to be raised on 404
We've initially made Down::NotFound a catch-all to have easier
integration in Shrine, as there we wanted to convert any Down exceptions
into validation errors. However, Down::NotFound being a subclass to
almost every exception this is neither intuitive nor useful.

In fact, in Shrine 3.0 we need a Down exception that means "404" for
remote storages to be able to detect missing files and raise
Shrine::FileNotFound. So, we repurpose Down::NotFound to mean "404" and
be a descendant of Down::ClientError.

This is a breaking change, so we'll be bumping the major version.