Fix keyword arguments warnings on Ruby 2.7
Ruby 2.7 adds deprecation warnings that guide towards separation of
positional and keyword arguments that's planned for Ruby 3.0. See the
following article for more details:

https://www.ruby-lang.org/en/news/2019/12/12/separation-of-positional-and-keyword-arguments-in-ruby-3-0/

We fix the warnings, making things works both on Ruby 2.7 and older
versions.

In the process, we deprecate mixing headers with other options for the
NetHttp backend, as making it work across Ruby versions is sufficiently
complex. This was supported originally because open-uri accepts headers
like that, but we've since introduced the :headers option which is
preferred now.