Deprecate the wget backend
It doesn't appear that anyone is using it, since I haven't received a
single bug report. And it's a maintenance burden, e.g. I cannot install
posix-spawn on my M1 on Ruby 3.3.1 anymore, and the tests without
posix-spawn are failing for some reason. So, we remove it from the
docs and recommend one of the three other backends.

This will allow us to potentially use WebMock in the future instead of
spinning up a real HTTP server.